requirements.txt
pydantic
python-dotenv
ipykernel
langchain
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence-transformers
langchain_huggingface
faiss-cpu
beautifulsoup4
streamlit
langchain-ollama
langchain_groq
fastapi
uvicorn